Drop glue/sop.py, as the sop module is now in PyPi
I've published the `sop` module on PyPi: https://pypi.org/project/sop/

So we do not need to ship it in this repository.

As long as it's installed cleanly in the local system (e.g. with "pip
install"; or in debian, see https://bugs.debian.org/943742 for the
ITP, https://ftp-master.debian.org/new/python-sop_0.1.3-1.html for its
place in the NEW queue), then sopgpy should work OK.